app/authentication/schemas/auth_schema.py
from typing import Optional
from pydantic import BaseModel, validator

class Token(BaseModel):
    access_token: str
    token_type: str


class TokenData(BaseModel):
    username: Optional[str] = None

class EmailVerification(BaseModel):
    token: str

class PasswordReset(BaseModel):
    token: str
    password: str
    confirm_password: str

    @validator('confirm_password')
    def passwords_match(cls, v, values, **kwargs):
        if 'password' in values and v != values['password']:
            raise ValueError('passwords do not match')
        return v